requirements.txt
agno==1.1.4
fastapi==0.115.8
uvicorn==0.34.0
groq==0.18.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.6
SQLAlchemy==2.0.38
peewee==3.17.9
tqdm==4.67.1
lancedb==0.19.0
sentence-transformers==3.4.1
speechrecognition==3.14.1
pyttsx3==2.98
pyaudio==0.2.14
huggingface-hub==0.29.1
python-multipart==0.0.20
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.5.2
tantivy==0.22.0
tokenizers==0.21.0
